// src/enums/user-type.enum.ts
export enum ProfileTypeEnum {
  ADMIN = "ADMIN",
  MEMBER = "MEMBER",
  GUEST = "GUEST",
  SUPER_ADMIN = "SUPER_ADMIN",
  DEPARTMENT_HEAD = "DEPARTMENT_HEAD",
  MINISTER_IN_CHARGE = "MINISTER_IN_CHARGE",
  UNIT_HEAD = "UNIT_HEAD",
  SYSTEM_SUPPORT = "SYSTEM_SUPPORT",
  ADMIN_STAFF = "ADMIN_STAFF",
  PASTOR = "PASTOR",
  WORKER = "WORKER",
}

export enum UserStatusEnum {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  MOVED = "MOVED",
  ARCHIVED = "ARCHIVED",
}

export enum DeviceTypeEnum {
  WEB = "WEB",
  APP = "APP",
}

export enum VerificationTypeEnum {
  EMAIL = "EMAIL",
  PHONE = "PHONE",
}

export enum GenderEnum {
  MALE = "MALE",
  FEMALE = "FEMALE",
}

export enum AgeRangeEnum {
  CHILD = "0-12",
  TEEN = "13-19",
  ADULT = "20-64",
  SENIOR = "65-100",
}

export enum MaritalStatusEnum {
  SINGLE = "SINGLE",
  MARRIED = "MARRIED",
  DIVORCED = "DIVORCED",
  WIDOWED = "WIDOWED",
}